Add capabilities for a shared folder to the API

+++ packages/api/cms-api/src/dam/files/folders.resolver.ts
@@ -1,14 +1,21 @@
-import { NotFoundException, Type } from "@nestjs/common";
+import { InjectRepository } from "@mikro-orm/nestjs";
+import { EntityRepository } from "@mikro-orm/postgresql";
+import { ForbiddenException, Inject, NotFoundException, Type } from "@nestjs/common";
 import { Args, ID, Mutation, ObjectType, Parent, Query, ResolveField, Resolver } from "@nestjs/graphql";
 
+import { GetCurrentUser } from "../../auth/decorators/get-current-user.decorator";
 import { SkipBuild } from "../../builds/skip-build.decorator";
 import { PaginatedResponseFactory } from "../../common/pagination/paginated-response.factory";
 import { AffectedEntity } from "../../user-permissions/decorators/affected-entity.decorator";
 import { RequiredPermission } from "../../user-permissions/decorators/required-permission.decorator";
+import { CurrentUser } from "../../user-permissions/dto/current-user";
+import { ACCESS_CONTROL_SERVICE } from "../../user-permissions/user-permissions.constants";
+import { AccessControlServiceInterface } from "../../user-permissions/user-permissions.types";
 import { DamScopeInterface } from "../types";
 import { EmptyDamScope } from "./dto/empty-dam-scope";
 import { createFolderArgs, createFolderByNameAndParentIdArgs, FolderArgsInterface, FolderByNameAndParentIdArgsInterface } from "./dto/folder.args";
 import { CreateFolderInput, UpdateFolderInput } from "./dto/folder.input";
+import { SHARED_FOLDER_NAME } from "./entities/folder.constants";
 import { FolderInterface } from "./entities/folder.entity";
 import { FoldersService } from "./folders.service";
 
@@ -39,7 +46,11 @@ export function createFoldersResolver({
     @RequiredPermission(["dam"], { skipScopeCheck: !hasNonEmptyScope })
     @Resolver(() => Folder)
     class FoldersResolver {
-        constructor(private readonly foldersService: FoldersService) {}
+        constructor(
+            private readonly foldersService: FoldersService,
+            @Inject(ACCESS_CONTROL_SERVICE) private accessControlService: AccessControlServiceInterface,
+            @InjectRepository("DamFolder") private readonly foldersRepository: EntityRepository<FolderInterface>,
+        ) {}
 
         @Query(() => [Folder])
         async damFoldersFlat(
@@ -55,13 +66,22 @@ export function createFoldersResolver({
         }
 
         @Query(() => Folder)
-        @AffectedEntity(Folder)
-        async damFolder(@Args("id", { type: () => ID }) id: string): Promise<FolderInterface> {
+        @RequiredPermission(["dam"], { skipScopeCheck: true }) // Custom scope check to handle the shared folder's fake scope
+        async damFolder(@Args("id", { type: () => ID }) id: string, @GetCurrentUser() user: CurrentUser): Promise<FolderInterface> {
             const folder = await this.foldersService.findOneById(id);
             if (!folder) {
                 throw new NotFoundException(id);
             }
-            return folder;
+
+            if (folder.scope) {
+                if (this.accessControlService.isAllowed(user, "dam", folder.scope) || folder.isSharedBetweenAllScopes) {
+                    return folder;
+                } else {
+                    throw new ForbiddenException();
+                }
+            } else {
+                return folder;
+            }
         }
 
         @Query(() => Folder, { nullable: true })
@@ -77,7 +97,15 @@ export function createFoldersResolver({
             @Args("input", { type: () => CreateFolderInput }) input: CreateFolderInput,
             @Args("scope", { type: () => Scope, defaultValue: hasNonEmptyScope ? undefined : {} }) scope: typeof Scope,
         ): Promise<FolderInterface> {
-            return this.foldersService.create(input, nonEmptyScopeOrNothing(scope));
+            if (input.name === SHARED_FOLDER_NAME) {
+                if (nonEmptyScopeOrNothing(scope) === undefined) {
+                    throw new Error("Can't create a shared folder if the DAM isn't scoped");
+                }
+
+                return this.foldersService.createSharedFolder({ scope: nonEmptyScopeOrNothing(scope) });
+            }
+
+            return this.foldersService.create({ ...input }, nonEmptyScopeOrNothing(scope));
         }
 
         @Mutation(() => Folder)
@@ -87,6 +115,11 @@ export function createFoldersResolver({
             @Args("id", { type: () => ID }) id: string,
             @Args("input", { type: () => UpdateFolderInput }) input: UpdateFolderInput,
         ): Promise<FolderInterface> {
+            const folder = await this.foldersRepository.findOneOrFail({ id });
+            if (folder?.isSharedBetweenAllScopes) {
+                throw new Error("Can't edit shared folder");
+            }
+
             return this.foldersService.updateById(id, input);
         }
 
@@ -97,6 +130,11 @@ export function createFoldersResolver({
             @Args("targetFolderId", { type: () => ID, nullable: true }) targetFolderId: string,
             @Args("scope", { type: () => Scope, defaultValue: hasNonEmptyScope ? undefined : {} }) scope: typeof Scope,
         ): Promise<FolderInterface[]> {
+            const sharedFolder = await this.foldersRepository.findOneOrFail({ isSharedBetweenAllScopes: true });
+            if (folderIds.includes(sharedFolder.id)) {
+                throw new Error("Can't move shared folder");
+            }
+
             return this.foldersService.moveBatch({ folderIds, targetFolderId }, nonEmptyScopeOrNothing(scope));
         }
 
@@ -104,6 +142,11 @@ export function createFoldersResolver({
         @AffectedEntity(Folder)
         @SkipBuild()
         async deleteDamFolder(@Args("id", { type: () => ID }) id: string): Promise<boolean> {
+            const folder = await this.foldersRepository.findOneOrFail({ id });
+            if (folder?.isSharedBetweenAllScopes) {
+                throw new Error("Can't delete shared folder");
+            }
+
             return this.foldersService.delete(id);
         }
